requirements.txt
boto
flask
kafka-python
